move RoutingChip to chips module
