+++ src/modules/routing/index.js
@@ -1,10 +1,5 @@
 import './i18n';
-import { RoutingChip } from './components/assistChip/RoutingChip';
 import { ProposalContextContent } from './components/contextMenu/ProposalContextContent';
-
-if (!window.customElements.get(RoutingChip.tag)) {
-	window.customElements.define(RoutingChip.tag, RoutingChip);
-}
 
 if (!window.customElements.get(ProposalContextContent.tag)) {
 	window.customElements.define(ProposalContextContent.tag, ProposalContextContent);
